Create a load_bboxes module
