+++ tests/test_unit/test_load_bboxes.py
@@ -1,0 +1,1 @@
+"""Test suite for the load_bboxes module."""
